完成了照片删除的善后功能，以及用户信息的修改

+++ DTO/src/main/java/com/seek/food/dto/User/UserDTO.java
@@ -15,7 +15,7 @@ public class UserDTO {
     private String username;
     private String phoneNumber;
     private String password;
-    private Byte sex;
+    private Integer sex;
     private String headerImageAddr;
     private LocalDate birthday;
     private LocalDateTime createTime;
